Move functions making API calls into their own file
