+++ src/Components/MessageList.jsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { List } from 'semantic-ui-react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { fetchMessagesForChannel } from '../Actions/index';
+import SLACK_API from '../Utils/Api';
 import Message from './Message';
 import owl from '../images/avatars/owl.png';
 import type { Dispatch, OwnProps, State } from '../FlowTypes/';
@@ -67,6 +67,7 @@ export const mapStateToProps = (state: State, ownProps: OwnProps) => {
   };
 };
 
+const fetchMessagesForChannel = SLACK_API.fetchMessagesForChannel;
 export const mapDispatchToProps = (dispatch: Dispatch) =>
   bindActionCreators({ fetchMessagesForChannel }, dispatch);
