+++ src/Components/Toolbar.jsx
@@ -5,7 +5,8 @@ import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { Dropdown, Image, Menu } from 'semantic-ui-react';
-import { selectChannel, fetchChannels } from '../Actions/index';
+import { selectChannel } from '../Actions/index';
+import SLACK_API from '../Utils/Api'
 import convertScoreToColorAndEmoji from '../Utils/';
 import frustrated from '../images/emojis/frustrated.jpg';
 import happy from '../images/emojis/happy.jpg';
@@ -97,6 +98,7 @@ export const mapStateToProps = (state: State, ownProps: OwnProps) => {
   }
 };
 
+const fetchChannels = SLACK_API.fetchChannels;
 export const mapDispatchToProps = (dispatch: Dispatch) =>
 bindActionCreators({
   selectChannel, fetchChannels,
